Create a dedicated class SliceRule so we can add custom features like ignoreDependency(..)
