+++ archunit/src/main/java/com/tngtech/archunit/library/dependencies/DependencyRules.java
@@ -15,54 +15,14 @@
  */
 package com.tngtech.archunit.library.dependencies;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.TreeSet;
-
-import com.google.common.base.Joiner;
-import com.tngtech.archunit.core.domain.Dependency;
 import com.tngtech.archunit.lang.ArchCondition;
-import com.tngtech.archunit.lang.ArchRule;
-import com.tngtech.archunit.lang.ConditionEvents;
-import com.tngtech.archunit.lang.SimpleConditionEvent;
-
-import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.all;
 
 class DependencyRules {
     static ArchCondition<Slice> beFreeOfCycles() {
         return new SliceCycleArchCondition();
     }
 
-    static ArchRule slicesShouldNotDependOnEachOtherIn(Slices.Transformer inputTransformer) {
-        return all(inputTransformer).should(notDependOnEachOther(inputTransformer));
-    }
-
-    private static ArchCondition<Slice> notDependOnEachOther(final Slices.Transformer inputTransformer) {
-        return new ArchCondition<Slice>("not depend on each other") {
-            @Override
-            public void check(Slice slice, ConditionEvents events) {
-                Slices dependencySlices = inputTransformer.transform(slice.getDependencies());
-                for (Slice dependencySlice : dependencySlices) {
-                    events.add(SimpleConditionEvent.violated(slice, describe(slice, dependencySlice)));
-                }
-            }
-
-            private String describe(Slice slice, Slice dependencySlice) {
-                return String.format("%s calls %s:%n%s",
-                        slice.getDescription(),
-                        dependencySlice.getDescription(),
-                        joinDependencies(slice, dependencySlice));
-            }
-
-            private String joinDependencies(Slice from, Slice to) {
-                List<String> parts = new ArrayList<>();
-                for (Dependency dependency : new TreeSet<>(from.getDependencies())) {
-                    if (to.contains(dependency.getTargetClass())) {
-                        parts.add(dependency.getDescription());
-                    }
-                }
-                return Joiner.on(System.lineSeparator()).join(parts);
-            }
-        };
+    static SliceRule slicesShouldNotDependOnEachOtherIn(Slices.Transformer inputTransformer) {
+        return new SliceRule(inputTransformer);
     }
 }
